Add the support of the multiple locale languages

+++ ui/plugins/cpp.ts
@@ -366,7 +366,7 @@ export function cppPlugin(options: CppPluginOptions = {}): Plugin {
             totalCompressedSize += compressed.length;
 
             header += `// ${asset.path}\n`;
-            header += `const uint16_t ${compressed.name}_L = ${compressed.length};\n`;
+            header += `const uint32_t ${compressed.name}_L = ${compressed.length};\n`;
             header += `const uint8_t ${compressed.name}[] PROGMEM = {\n${compressed.array}\n};\n\n`;
             header += `inline void serve${pascalCase(asset.name)}(AsyncWebServerRequest* request) {\n`;
             header += `  AsyncWebServerResponse *response = request->beginResponse_P(200, "${asset.contentType}", ${compressed.name}, ${compressed.name}_L);\n`;
